moved reusable components and added optional extra classnames
