+++ stockflux-securities-master/src/components/TextField.js
@@ -1,13 +1,15 @@
 import React from "react";
 import "./TextField.css";
 
-const TextField = ({ value, id, onChange, readOnly }) => {
+const TextField = ({ value, id, onChange, readOnly, className }) => {
   return (
     <input
       value={value}
       id={id}
       readOnly={readOnly}
-      className={"input-form-input" + (readOnly ? " read-only" : "")}
+      className={
+        "input-form-input" + (readOnly ? " read-only" : "") + ` ${className}`
+      }
       onChange={onChange}
     />
   );
