+++ stockflux-securities-master/src/components/ToggleSwitch.css
