+++ stockflux-securities-master/src/components/ToggleSwitch.js
@@ -1,19 +1,19 @@
 import React from "react";
 import "./ToggleSwitch.css";
 
-const ToggleSwitch = ({ value, id, checked, onChange }) => {
+const ToggleSwitch = ({ value, id, checked, onChange, className }) => {
   return (
-    <div>
+    <>
       <input
         value={value}
         id={id}
         checked={checked}
         type="checkbox"
-        className="input-form-toggle"
+        className={"input-form-toggle" + ` ${className}`}
         onChange={onChange}
       />
       <label className="toggle-switch" htmlFor={id} />
-    </div>
+    </>
   );
 };
